+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 2. [](#improved)
     * Improved `Filesystem::basename()` and `Filesystem::pathinfo()` to be unicode-safe
     * Made path handling unicode-safe, use new `Utils::basename()` and `Utils::pathinfo()` everywhere
+3. [](#bugfix)
+    * Fixed error on thumbnail image creation
     * Fixed MimeType for `gzip` (`application/x-gzip`)
 
 # v1.7.28